avoid crashes in statical linked binaries when dlopen()
Even in current (1.0.55) releaes of uClibc-ng there will be 100% reproducible crashes of statically linked binaries (on all kind of platforms), when calling "dlopen(...)" with wrong or non-existing .so-files).

#0  0x0000000000404b62 in _dl_load_shared_library ()
#1  0x0000000000404d49 in do_dlopen ()
#2  0x0000000000405286 in dlopen ()

This is caused by missing checks on "_dl_loaded_modules" in "ldso/ldso/dl-elf.c".
When "_dl_loaded_modules" is NULL in static linked binaries, it becomes dereferenced and causes an segfault.

This patch fixes the issue by adding an extra assignment-check for "_dl_loaded_modules".

Signed-off-by: Stephan Baerwolf <stephan@matrixstorm.com>